# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Jewellery management API — retail, wholesale, GST, Girvi, ledger, reports
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/daybook:
    get:
      operationId: getDaybook
      tags: [reports]
      parameters:
        - in: query
          name: date
          schema: { type: string, format: date }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DaybookReport" }
...
  /reports/gst:
    get:
      operationId: getGstReport
      tags: [reports]
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GstReport" }
...
components:
  schemas:
...
    GstSlabRow:
      type: object
      properties:
        gstRate: { type: number }
        taxableAmount: { type: number }
        cgst: { type: number }
        sgst: { type: number }
        totalTax: { type: number }
        invoiceCount: { type: integer }
      required: [gstRate, taxableAmount, cgst, sgst, totalTax, invoiceCount]
    GstReport:
      type: object
      properties:
        from: { type: string, format: date }
        to: { type: string, format: date }
        totalTaxable: { type: number }
        totalCgst: { type: number }
        totalSgst: { type: number }
        totalTax: { type: number }
        slabs:
          type: array
          items: { $ref: "#/components/schemas/GstSlabRow" }
      required: [from, to, totalTaxable, totalCgst, totalSgst, totalTax, slabs]
...
    DaybookEntry:
      type: object
      properties:
        time: { type: string, format: date-time }
        kind: { type: string, enum: [sale, payment_in, payment_out, girvi_loan, girvi_payment, repair_collected, scheme_installment] }
        reference: { type: string }
        party: { type: string }
        amount: { type: number }
        mode: { type: string }
      required: [time, kind, reference, party, amount, mode]
    DaybookReport:
      type: object
      properties:
        date: { type: string, format: date }
        openingCash: { type: number }
        cashIn: { type: number }
        cashOut: { type: number }
        upiIn: { type: number }
        cardIn: { type: number }
        bankIn: { type: number }
        netSales: { type: number }
        entries:
          type: array
          items: { $ref: "#/components/schemas/DaybookEntry" }
      required: [date, openingCash, cashIn, cashOut, upiIn, cardIn, bankIn, netSales, entries]