# lib/api-zod/src/generated/types/daybookEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jewellery management API — retail, wholesale, GST, Girvi, ledger, reports
 * OpenAPI spec version: 0.1.0
 */
import type { DaybookEntryKind } from "./daybookEntryKind";

export interface DaybookEntry {
  time: Date;
  kind: DaybookEntryKind;
  reference: string;
  party: string;
  amount: number;
  mode: string;
}